[CONTRIB] TYPO3 10 compatibility
[TASK] Retrieve extension configuration since TYPO3 v9.0
[TASK] Remove obsolete TCA configurations since TYPO3 v9.5 and v10.4
[TASK] Fix injection for classes
Fix syntax for inject-method-names and instanciate missing class
on initialize action.
[TASK] Rewrite TYPO3 VersionNumberUtility for compatibility
[TASK] Rewrite @injection and @ignore annotations for compatibility
[TASK] Compatibility for sending author mail on approve via backend
This change replaces eid with middleware request.
[TASK] Use TemplateService instead of TSFE to fetch typoscript in BE
Typoscript for evaluation of mail notification, as introduced with 6d1bbf7
is now fetched with TemplateService cause EidUtility is obsolete
and TypoScriptFrontendController returns first root-page setup only.
This fixes an error when unhide comment entry in backend,
cause of Hook ProcessDatamap.php.
There will still be error when unhide comment entry in backend,
if typoscript "sendMailToAuthorAfterPublish" is true.
This references to TYPO3v10.4 compatibility